aw-models: Moved tests to respective files
