+++ aw-models/src/lib.rs
@@ -1,21 +1,8 @@
 extern crate serde;
-extern crate serde_json;
+#[macro_use] extern crate serde_json;
 #[macro_use] extern crate serde_derive;
 extern crate chrono;
 #[macro_use] extern crate log;
-
-mod duration;
-mod bucket;
-mod event;
-mod timeinterval;
-mod query;
-
-pub use self::bucket::Bucket;
-pub use self::bucket::BucketMetadata;
-pub use self::bucket::BucketsExport;
-pub use self::event::Event;
-pub use self::timeinterval::TimeInterval;
-pub use self::query::Query;
 
 // TODO: Move me to an appropriate place
 #[macro_export]
@@ -32,3 +19,15 @@ macro_rules! json_map {
     }};
 }
 
+mod duration;
+mod bucket;
+mod event;
+mod timeinterval;
+mod query;
+
+pub use self::bucket::Bucket;
+pub use self::bucket::BucketMetadata;
+pub use self::bucket::BucketsExport;
+pub use self::event::Event;
+pub use self::timeinterval::TimeInterval;
+pub use self::query::Query;
