+++ aw-models/src/timeinterval.rs
@@ -93,3 +93,24 @@ impl<'de> Deserialize<'de> for TimeInterval {
         deserializer.deserialize_str(TimeIntervalVisitor)
     }
 }
+
+#[test]
+fn test_timeinterval() {
+    use std::str::FromStr;
+
+    let start = DateTime::from_str("2000-01-01T00:00:00Z").unwrap();
+    let end = DateTime::from_str("2000-01-02T00:00:00Z").unwrap();
+    let period_str = "2000-01-01T00:00:00+00:00/2000-01-02T00:00:00+00:00";
+    let duration = end - start;
+    let tp = TimeInterval::new(start, end);
+    assert_eq!(tp.start(), &start);
+    assert_eq!(tp.end(), &end);
+    assert_eq!(tp.duration(), duration);
+    assert_eq!(tp.to_string(), period_str);
+
+    let tp = TimeInterval::new_from_string(period_str).unwrap();
+    assert_eq!(tp.start(), &start);
+    assert_eq!(tp.end(), &end);
+    assert_eq!(tp.duration(), duration);
+    assert_eq!(tp.to_string(), period_str);
+}
